initial project and create Structure
I created a new project and started it from scratch. I cleaned the starter project of the files I didn't need, put together a small application structure and made the initial markup. Added styles and basic components. I prepared the project for further work and added animations to the background.