Added: Blocks Completed and Failed
I added new completed and failed blocks to the main section, styled them, set up an adaptive one, and now I can work on the site's functionality. But for now, I'll put this idea on hold a little bit, as there is also a thought about adding a footer block. And yes, as an adaptive, I chose a rubber layout using grids to control the completed and failed blocks.