Updated: page layout and styles
As it turned out, I misunderstood the tk and needed to add not one task, but many, so the design and structure of the project had to be redone. Also, with the redesign of the structure, I updated the styles a little, moved the general styles to separate files, and moved all the styles that are in the main block to a separate folder from where I have already connected everything.