Corrections: Structure and Styles
The application's appearance has been finalized (rubber layout) and work on the application structure has been completed. A new assets folder has been added, which contains all the styles. They are just as structurally broken down as the components. I haven't figured out the relative paths yet, but it's not that important anymore, the main thing is that it works.